Adds an item route
